+++ tests/testthat/test-provider-openai.R
@@ -1,14 +1,14 @@
 # Getting started --------------------------------------------------------
 
 test_that("can make simple request", {
-  chat <- chat_openai("Be as terse as possible; no punctuation")
+  chat <- chat_openai("Be as terse as possible; no punctuation", model = "gpt-4o-mini")
   resp <- chat$chat("What is 1 + 1?", echo = FALSE)
   expect_match(resp, "2")
   expect_equal(chat$last_turn()@tokens, c(27, 1))
 })
 
 test_that("can make simple streaming request", {
-  chat <- chat_openai("Be as terse as possible; no punctuation")
+  chat <- chat_openai("Be as terse as possible; no punctuation", model = "gpt-4o-mini")
   resp <- coro::collect(chat$stream("What is 1 + 1?"))
   expect_match(paste0(unlist(resp), collapse = ""), "2")
 })
@@ -20,14 +20,14 @@ test_that("defaults are reported", {
 })
 
 test_that("respects turns interface", {
-  chat_fun <- chat_openai
+  chat_fun <- function(...) chat_openai(..., model = "gpt-4o-mini")
 
   test_turns_system(chat_fun)
   test_turns_existing(chat_fun)
 })
 
 test_that("all tool variations work", {
-  chat_fun <- chat_openai
+  chat_fun <- function(...) chat_openai(..., model = "gpt-4o-mini")
 
   test_tools_simple(chat_fun)
   test_tools_async(chat_fun)
@@ -36,13 +36,13 @@ test_that("all tool variations work", {
 })
 
 test_that("can extract data", {
-  chat_fun <- chat_openai
+  chat_fun <- function(...) chat_openai(..., model = "gpt-4o-mini")
 
   test_data_extraction(chat_fun)
 })
 
 test_that("can use images", {
-  chat_fun <- chat_openai
+  chat_fun <- function(...) chat_openai(..., model = "gpt-4o-mini")
 
   test_images_inline(chat_fun)
   test_images_remote(chat_fun)